fix!: remove to simplistic minification
the minification was based on simple replacements and the assumption
that mostly all whitespace can just be removed.
But at least for pre tags this is just not true.
To avoid such unintended behaviour this bad abstraction is no longet
supported. We keep the function for the sake of testing though.

BREAKING CHANGE: remove rendering option `minify`